refactor: rename P2 violations in yaml, basemap, spreadsheet, folder plugins (#1213)
- yaml-plugin: YamlEntity.ts → yamlEntityTypes.ts
- basemap-plugin: viewportValidation.ts → viewportValidationUtils.ts, mapStyle.ts → mapStyleUtils.ts
- spreadsheet-plugin: hash.ts → hashUtils.ts
- folder-plugin: folderValidation.ts → folderValidationUtils.ts, metadata.ts → metadataConstants.ts

Rollback: git revert <hash> --no-edit
