+++ plugins/basemap-plugin/src/common/shared/viewportValidationUtils.ts
