+++ plugins/basemap-plugin/src/ui/utils/mapStyleUtils.ts
