+++ plugins/folder-plugin/src/common/shared/folderValidationUtils.ts
