+++ plugins/spreadsheet-plugin/src/services/SpreadsheetTabularApiDriver.ts
@@ -22,7 +22,7 @@ import { toNodeId } from '@hierarchidb/core-types';
 import { getDBName } from '@hierarchidb/util';
 import { SpreadsheetMetadataManager } from './SpreadsheetMetadataManager.js';
 import { SpreadsheetStorePort } from './SpreadsheetStorePort.js';
-import { hashFile } from './utils/hash.js';
+import { hashFile } from './utils/hashUtils.js';
 import { matchesFilters, normalizeValueForResult, prepareFilters, type PreparedFilter, type TabularRow } from './utils/filtering.js';
 import { SPREADSHEET_PLUGIN_ID } from '../common/constants.js';
