+++ plugins/spreadsheet-plugin/src/services/utils/hashUtils.ts
