+++ plugins/yaml-plugin/src/common/types/yamlEntityTypes.ts
